Changed json.org to json-simple

+++ src/org/vatvit/irccloudandroid/ChannelActivity.java
@@ -42,7 +42,7 @@ public class ChannelActivity extends Activity {
 		client = app.getClient();
 		Bundle extras = getIntent().getExtras();
 
-		int cid = extras.getInt("cid");
+		long cid = extras.getLong("cid");
 		for (Server serv : client.getServers()) {
 			if (serv.getCid() == cid) {
 				server = serv;
